apps/api-django/pyproject.toml
[project]
name = "atlas-api"
version = "0.1.0"
description = "The central API Gateway"
authors = [
    {name = "Jason Booth"}
]
requires-python = "^3.10"
dependencies = [
    "django (>=5.1,<6.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "markdown (>=3.10.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)"
]
